backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
pymongo==4.6.0
mongoengine==0.27.0
djongo==1.3.6
requests
dotenv

